Adding configurationDirectory and configurationString types (#150)
  * Configuration() now returns the Config interface to accommodate the different types that could be returned (e.g., configurationDirectory, configurationString)
