+++ internal/teststep/config_test.go
@@ -4,7 +4,6 @@
 package teststep
 
 import (
-	"context"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
@@ -56,118 +55,6 @@ func TestConfigurationRequest_Validate(t *testing.T) {
 				if diff := cmp.Diff(err.Error(), testCase.expectedError); diff != "" {
 					t.Errorf("expected error %s, got error %s", testCase.expectedError, err)
 				}
-			}
-		})
-	}
-}
-
-func TestConfiguration_HasProviderBlock(t *testing.T) {
-	t.Parallel()
-
-	testCases := map[string]struct {
-		config   configuration
-		expected bool
-	}{
-		"no-config": {
-			config:   configuration{},
-			expected: false,
-		},
-		"provider-meta-attribute": {
-			config: configuration{
-				raw: `
-resource "test_test" "test" {
-  provider = test.test
-}
-`,
-			},
-			expected: false,
-		},
-		"provider-object-attribute": {
-			config: configuration{
-				raw: `
-resource "test_test" "test" {
-  test = {
-	provider = {
-	  test = true
-	}
-  }
-}
-`,
-			},
-			expected: false,
-		},
-		"provider-string-attribute": {
-			config: configuration{
-				raw: `
-resource "test_test" "test" {
-  test = {
-	provider = "test"
-  }
-}
-`,
-			},
-			expected: false,
-		},
-		"provider-block-quoted-with-attributes": {
-			config: configuration{
-				raw: `
-provider "test" {
-  test = true
-}
-
-resource "test_test" "test" {}
-`,
-			},
-			expected: true,
-		},
-		"provider-block-unquoted-with-attributes": {
-			config: configuration{
-				raw: `
-provider test {
-  test = true
-}
-
-resource "test_test" "test" {}
-`,
-			},
-			expected: true,
-		},
-		"provider-block-quoted-without-attributes": {
-			config: configuration{
-				raw: `
-provider "test" {}
-
-resource "test_test" "test" {}
-`,
-			},
-			expected: true,
-		},
-		"provider-block-unquoted-without-attributes": {
-			config: configuration{
-				raw: `
-provider test {}
-
-resource "test_test" "test" {}
-`,
-			},
-			expected: true,
-		},
-	}
-
-	for name, testCase := range testCases {
-		name, testCase := name, testCase
-
-		t.Run(name, func(t *testing.T) {
-			t.Parallel()
-
-			got, err := testCase.config.HasProviderBlock(context.Background())
-
-			if err != nil {
-				t.Errorf("unexpected error: %s", err)
-			}
-
-			if testCase.expected != got {
-				t.Errorf("expected %t, got %t", testCase.expected, got)
 			}
 		})
 	}
